Move ByteSize to toml module removing bytes module
Move code for ByteSize from the bytes.rs to toml.rs and then remove the
bytes module entirely.
